Replaced inherited MongoDB with MySQL. All tests but CRUD and server still pass.
