+++ package.json
@@ -10,10 +10,7 @@
     "build_client": "webpack",
     "preseed": "node database-mongodb/unsplashAPIcalls.js",
     "seed": "node database-mongodb/initiateSeeding.js",
-    "seedMySQL": "node database-mongodb/MySQL/initiateSeeding.js",
-    "seedRiak": "node database-mongodb/Riak/initiateSeeding.js",
-    "dbCompareMySQL": "node database-mongodb/MySQL/testDatabaseQueryEfficiency.js",
-    "dbCompareRiak": "node database-mongodb/Riak/testDatabaseQueryEfficiency.js"
+    "dbEfficiency": "node database-mongodb/testDatabaseQueryEfficiency.js"
   },
   "author": "",
   "license": "ISC",
